Moved all discovery to a new pass
